Kafka/Consumer/requirements.txt
kafka-python
aiokafka
flask-kafka
Flask==2.0.1
flask-cors==3.0.6
confluent_kafka
